Polish the settings window and fix what review turned up
Three defects, all found by reviewing the branch rather than running it:

Alt+F4 while recording a shortcut was captured instead of closing the window. The
recorder marked every key handled before deciding what it was, and Alt+F4 reaches
an app as an ordinary key — it is DefWindowProc that turns it into a close. So the
window stayed open and Alt+F4 was accepted as the binding, registering it globally:
it would then step brightness rather than close windows in every application, and
the setting persisted across restarts.

A press during a rescan claimed no display was connected. RefreshAsync empties
Displays until every brightness read finishes, and the rescan triggers on unlock,
resume, display changes and USB arrival — exactly when the brightness keys get
used.

The hotkey hook could throw out of the window procedure: IntPtr.ToInt32 is checked,
and WM_HOTKEY sits below WM_USER, so any process can post one with a wParam outside
Int32, as Windows itself does for IDHOT_SNAPWINDOW.

The rest is presentation. The tabs now line up with the title bar icon and carry a
wider selection indicator; the Shortcuts page gets a wider column, left-aligned
footnotes and a note that a press is worth 10%; and the second tab is called
Support, after the donate and repository links that are its reason to exist.

The README's shortcut section goes — the landing page mentions them alongside the
tray window, which is the same idea — and the comments this branch added are cut
back to the house style, keeping the framework details that get re-broken
otherwise.

+++ src/app/Services/HotKeyService.cs
@@ -32,17 +32,14 @@ public interface IHotKeyCoordinator
     // accepted.
     HotKeyApplyResult Apply(AppSettings settings);
 
-    // Releases the registrations while the user types a new combination — Windows
-    // delivers a registered shortcut as WM_HOTKEY, so it would otherwise never reach
-    // the settings window as a key press.
+    // Releases the registrations while the user types a new combination, which would
+    // otherwise arrive as WM_HOTKEY instead of a key press.
     void SetShortcutsSuspended(bool suspended);
 }
 
-// Registers the global brightness shortcuts against the main window's HWND and
-// turns WM_HOTKEY into events. The window is normally hidden in the tray, but its
-// handle stays alive and keeps pumping messages, so the shortcuts keep working.
-// Every registration carries MOD_NOREPEAT: one press is one step, so holding a key
-// down can't race through the whole range.
+// Registers the global brightness shortcuts against the main window's HWND and turns
+// WM_HOTKEY into events. The window is normally hidden in the tray, but its handle
+// keeps pumping messages. MOD_NOREPEAT keeps one press to one step.
 internal sealed class HotKeyService : IDisposable
 {
     // Application-defined ids must be 0x0000-0xBFFF.
@@ -63,8 +60,7 @@ internal sealed class HotKeyService : IDisposable
     public HotKeyService(Window window)
         => _window = window ?? throw new ArgumentNullException(nameof(window));
 
-    // Re-registers both shortcuts from scratch. Safe to call repeatedly (every time
-    // settings change); previous registrations are always released first.
+    // Re-registers both shortcuts from scratch; previous ones are always released first.
     public HotKeyApplyResult Apply(bool enabled, HotKeyBinding? up, HotKeyBinding? down)
     {
         if (_disposed)
@@ -163,7 +159,9 @@ internal sealed class HotKeyService : IDisposable
             return IntPtr.Zero;
         }
 
-        switch (wParam.ToInt32())
+        // ToInt32 is checked, and any process can post WM_HOTKEY with an out-of-range
+        // wParam, which would throw out of the window procedure.
+        switch (wParam.ToInt64())
         {
             case IdBrightnessUp:
                 handled = true;
